Move NLG under apps
