+++ gramex/apps/admin2/gramex.yaml
@@ -82,66 +82,3 @@ url:
       headers:
         Content-Type: application/json      # Display as JSON
 
-  # NLG Webapp handlers
-  apps/admin2/nlg-data-selector-$*:
-    pattern: /$YAMLURL/initform
-    handler: FunctionHandler
-    kwargs:
-      headers:
-        Cache-Control: no-store
-      xsrf_cookies: false
-      function: gramex.nlg.webapp.init_form
-      redirect:
-        url: /$YAMLURL/nlg-editor
-  apps/admin2/nlg-editor-filehander-$*:
-    pattern: /$YAMLURL/nlg-editor
-    handler: FileHandler
-    kwargs:
-      path: $YAMLPATH/nlg/index.html
-      template: true
-      auth:
-        login_url: /$YAMLURL/login
-  apps/admin2/tablepreview-$*:
-    pattern: /$YAMLURL/preview/(.*)/(.*)
-    handler: FormHandler
-    kwargs:
-      url: $GRAMEXDATA/nlg/{_0}/{_1}
-      headers:
-        Cache-Control: no-store
-  apps/admin2/nlg-static_files-$*:
-    pattern: /$YAMLURL/nlg.js
-    handler: FileHandler
-    kwargs:
-      path: $YAMLPATH/nlg/js/nlg.js
-  apps/admin2/nlg-config-handler-$*:
-    pattern: /$YAMLURL/initconf
-    handler: FunctionHandler
-    kwargs:
-      function: gramex.nlg.webapp.get_init_config
-      headers:
-        Content-Type: application/json
-  apps/admin2/textproc-$*:
-    pattern: /$YAMLURL/textproc
-    handler: FunctionHandler
-    priority: 1000000
-    kwargs:
-      function: gramex.nlg.webapp.process_template
-      xsrf_cookies: false
-      headers:
-        Content-Type: application/json
-  apps/admin2/rendertmpl-$*:
-    pattern: /$YAMLURL/render-template
-    handler: FunctionHandler
-    kwargs:
-      function: gramex.nlg.webapp.render_template
-      xsrf_cookies: false
-      headers:
-        Content-Type: application/json
-  apps/admin2/nlg_gramopts-$*:
-    pattern: /$YAMLURL/get-gramopts
-    handler: FunctionHandler
-    kwargs:
-      function: gramex.nlg.webapp.get_gramopts
-      xsrf_cookies: false
-      headers:
-        Content-Type: application/json
