+++ gramex/apps/admin2/index.html
@@ -164,7 +164,7 @@
       <div class="h1 text-danger">Work in progress</div>
 
     {% elif current_tab == 'nlg' %}
-      {% include "nlg/grmform.html" %}
+      {% include "../nlg/grmform.html" %}
 
     {% end %}
   </div>
