+++ gramex/apps/nlg/narrative.py
