+++ gramex/apps/nlg/utils.py
